test: split routing suites around shared helpers
